requirements.txt
streamlit
pandas
requests
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv

streamlit
pandas
requests
python-dotenv

